+++ src/components/Header/Header.jsx
@@ -73,7 +73,9 @@ import { useNavigate } from "react-router-dom";
 
 const Header = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
+  const [firstName, setFirstName] = useState("");
   const navigate = useNavigate();
+  console.log(firstName)
 
   useEffect(() => {
     const handleClick = (e) => {
@@ -130,10 +132,19 @@ const Header = () => {
                 >
                   Find Caterers
                 </li>
+                {firstName && <li className={styles.booking}>
+                  Bookings
+                </li>}
                 <li className={styles.navContact}>+91 123456789</li>
-                <li className={styles.navLogin} onClick={openModal}>
+                {!firstName && <li className={styles.navLogin} onClick={openModal}>
                   Login/Register
-                </li>
+                </li>}
+                {firstName && <li className={styles.navProfile}>
+                {firstName}
+                </li>}
+                {firstName && <li className={styles.signout}>
+                SignOut
+                </li>}
               </ul>
             </div>
           </div>
@@ -145,13 +156,22 @@ const Header = () => {
           >
             Find Caterers
           </li>
+          {firstName && <li className={styles.booking}>
+                Bookings
+          </li>}
           <li className={styles.navContact}>+91 123456789</li>
-          <li className={styles.navLogin} onClick={openModal}>
-            Login/Register
-          </li>
+          {!firstName && <li className={styles.navLogin} onClick={openModal}>
+                  Login/Register
+                </li>}
+                {firstName && <li className={styles.navProfile}>
+                {firstName}
+                </li>}
+                {firstName && <li className={styles.signout}>
+                SignOut
+                </li>}
         </ul>
       </nav>
-      <LoginRegisterModal isOpen={isModalOpen} onClose={closeModal} />
+      <LoginRegisterModal firstName={firstName} setFirstName={setFirstName} isOpen={isModalOpen} onClose={closeModal} />
     </>
   );
 };
